优化代码模块
