+++ checksvn/build_app.bat
@@ -2,11 +2,11 @@
 SET CGO_ENABLED=0
 SET GOOS=windows
 SET GOARCH=amd64
-go build -o ./checksvn.exe
+go build -o ./bin/checksvn.exe
 
 :: linux
 SET CGO_ENABLED=0
 SET GOOS=linux
 SET GOARCH=amd64
-go build -o ./checksvn
+go build -o ./bin/checksvn
